+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/DebugTraceBlockStreamer.java
@@ -183,7 +183,9 @@ public class DebugTraceBlockStreamer {
 
             final boolean isOpcodeTracer = traceOptions.tracerType() == TracerType.OPCODE_TRACER;
 
-            for (final Transaction transaction : block.getBody().getTransactions()) {
+            final List<Transaction> transactions = block.getBody().getTransactions();
+            for (int i = 0; i < transactions.size(); i++) {
+              final Transaction transaction = transactions.get(i);
               if (isOpcodeTracer) {
                 streamOpcodeTransaction(
                     transaction,
@@ -198,6 +200,7 @@ public class DebugTraceBlockStreamer {
                       mapper.writeValueAsBytes(
                           buildTransactionResult(
                               transaction,
+                              i,
                               chainUpdater,
                               transactionProcessor,
                               protocolSpec,
@@ -249,10 +252,13 @@ public class DebugTraceBlockStreamer {
                   .getPreExecutionProcessor()
                   .createBlockHashLookup(blockchainQueries.getBlockchain(), header);
 
-          for (final Transaction transaction : block.getBody().getTransactions()) {
+          final List<Transaction> transactions = block.getBody().getTransactions();
+          for (int i = 0; i < transactions.size(); i++) {
+            final Transaction transaction = transactions.get(i);
             results.add(
                 buildTransactionResult(
                     transaction,
+                    i,
                     chainUpdater,
                     transactionProcessor,
                     protocolSpec,
@@ -320,6 +326,7 @@ public class DebugTraceBlockStreamer {
 
   private DebugTraceTransactionResult buildTransactionResult(
       final Transaction transaction,
+      final int transactionIndex,
       final TraceBlock.ChainUpdater chainUpdater,
       final MainnetTransactionProcessor transactionProcessor,
       final ProtocolSpec protocolSpec,
@@ -347,7 +354,8 @@ public class DebugTraceBlockStreamer {
             transaction,
             result,
             step.getOperationTracer().getTraceFrames(),
-            Optional.empty(),
+            Optional.of(block),
+            transactionIndex,
             accessListTracker.getTouchedAccounts());
 
     return step.buildResult(transactionTrace);