+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/DebugTraceTransactionStep.java
@@ -19,6 +19,7 @@ import org.hyperledger.besu.ethereum.api.jsonrpc.internal.results.DebugTraceTran
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.results.FourByteTracer;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.results.OpCodeLoggerTracerResult;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.results.calltrace.CallTracer;
+import org.hyperledger.besu.ethereum.api.jsonrpc.internal.results.calltrace.FlatCallTracer;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.results.tracing.diff.StateDiffTrace;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.results.tracing.diff.StateTraceGenerator;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.results.tracing.diff.StateTraceResult;
@@ -26,8 +27,6 @@ import org.hyperledger.besu.ethereum.debug.TraceOptions;
 import org.hyperledger.besu.ethereum.mainnet.ProtocolSpec;
 import org.hyperledger.besu.ethereum.vm.DebugOperationTracer;
 import org.hyperledger.besu.evm.tracing.OperationTracer;
-
-import com.fasterxml.jackson.annotation.JsonGetter;
 
 /**
  * Encapsulates both the {@link OperationTracer} and the logic to build a {@link
@@ -97,8 +96,7 @@ public interface DebugTraceTransactionStep {
             private final DebugOperationTracer tracer =
                 new DebugOperationTracer(traceOptions.opCodeTracerConfig(), recordChildCallGas);
             private final StateTraceGenerator generator = new StateTraceGenerator();
-            private final boolean diffMode =
-                Boolean.TRUE.equals(traceOptions.tracerConfig().getOrDefault("diffMode", false));
+            private final boolean diffMode = traceOptions.tracerConfigFlag("diffMode");
 
             @Override
             public OperationTracer getOperationTracer() {
@@ -134,8 +132,7 @@ public interface DebugTraceTransactionStep {
           };
       case FLAT_CALL_TRACER ->
           new DebugTraceTransactionStep() {
-            private final DebugOperationTracer tracer =
-                new DebugOperationTracer(traceOptions.opCodeTracerConfig(), recordChildCallGas);
+            private final FlatCallTracer tracer = new FlatCallTracer(traceOptions, protocolSpec);
 
             @Override
             public OperationTracer getOperationTracer() {
@@ -144,16 +141,9 @@ public interface DebugTraceTransactionStep {
 
             @Override
             public DebugTraceTransactionResult buildResult(final TransactionTrace trace) {
-              return new DebugTraceTransactionResult(trace, new UnimplementedTracerResult());
+              return new DebugTraceTransactionResult(trace, tracer.buildResult(trace));
             }
           };
     };
-  }
-
-  class UnimplementedTracerResult {
-    @JsonGetter("error")
-    public String getError() {
-      return "Not Yet Implemented";
-    }
   }
 }