+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/ExecuteTransactionStep.java
@@ -127,6 +127,7 @@ public class ExecuteTransactionStep implements Function<TransactionTrace, Transa
         result,
         traceFrames,
         transactionTrace.getBlock(),
+        transactionTrace.getTransactionIndex(),
         touchedAccounts);
   }
 }